Validate that we're not trying to do empty dispatches in FComputeShaderUtils::ValidateGroupCount.
RHI validation already contains this check, but the callstacks are not very useful when that fires, since the error gets caught on the RHI thread. Doing it in FComputeShaderUtils::ValidateGroupCount catches the problem at the correct spot, i.e. when the pass is added to RDG.

Also, don't Nanite material passes if there aren't any materials (since those would be empty dispatches).

#rnx
#jira UE-195648
#rb graham.wihlidal, christopher.waters

[CL 29012719 by mihnea balta in ue5-main branch]