AnimNext: Can now run standalone without anim blueprints for support.
Additions:
Basic port implementation - A simple class to allow native ports to be written and extended. Ideally ports should eventually be RigVM assets that allow arbitrary composition of operations.
IO 'terms' for schedule entries - Per-schedule allocated intermediate values that are mapped into schedule task's IO terms, allowing tasks to communicate via those temporaries. Ideally this should all be handled at the RigVM level when we change schedules to be graphs, but this works to get the system stood up for now.
Anim Next Mesh Component - A non-ticking mesh component that acts as a container for component-space transforms and performs no game-thread work.
Added profiling scopes to get a better picture of system performance.
Added some more built-in parameter handle types. This has highlighted the sheer amount of code that needs to be updated to add new types, so this needs a rework.
Added the ability for a parameter layer to be mapped from another parameter layer, to support task/term mappings.
Expanded FParamStackLayerHandle API to wrap more of FParamStackLayer.
Changed FScheduleTickFunction to run a slice of schedule instructions, and used this in the initialization path to ensure schedules run initially correctly in non-ticking editor worlds (this fixes editor thumbnails).
Included parent bone indices in TReferencePose to aid in component-space pose generation. Also added some helper functions in FGenerationTools to aid with this (should these be in another library?)

Fixes/changes:
Fixed AnimNext graph editing so that the correct graph is both edited by the graph editor AND compiled by the compiler.
Changed FOnLODRequiredBonesUpdateMulticast to be a thread-safe delegate as it is now called on worker threads.
Sanitised UObject-ptr value parameter access. Now the type has to request the T<UObject*> type (or T<TObjectPtr<>>) to work. Previously it allowed T<UObject> which was incorrect.
Fixed race condition in FDataRegistry by adding appropriate lock.
Added extra checking to higher-level APIs to deal with various scenarios where FScheduler calls could be made when worlds or world subsystems might be invalid (e.g. changing editor worlds).
Added a resolved hard object ptr to schedule entries that is only valid when schedules are running - this both provides GC safety and improves performance as the weak object ptr was being accessed for every adapter call.
Various extra null checks added to places where content can provide null values (parameter blocks, sequence player decorators etc.)
Fixed schedule editing to allow undo/redo.
Fixed FEvaluationVM to work if there is not a mesh present (as it tries to make a bone container, which can assert).
Removed the guard around not displaying object types in SRigVMGraphChangePinType so that wildcard pins can be set to object type values.
Changed a number of CVars to GetValueOnAnyThread as their calling functions are now called on worker threads for the first time.
Updated FRootMotionReset::ResetRootBoneForRootMotion to not be dependent on a bone container, allowing better root motion handling.
Moved some utility functions from editor to uncookedonly.

#rb Nicholas.Frechette, jaime.cifuentes

[CL 29013041 by thomas sarkanen in ue5-main branch]