Use GetCopyableFootprints for sizing the staging buffer in FD3D12Texture::Lock, rather than manually doing the math.
The math we were doing was wrong in some corner cases (such as when trying to read a PF_DepthStencil texture, because for legacy reasons GPixelFormats says that the block size for that format is 5 bytes), and the engine crashed because D3D returned E_INVALIDARG when we tried to execute the copy operation. While reading back a depth-stencil surface like that is almost never a good idea, we shouldn't crash when somebody does it. Anyway, using GetCopyableFootprints gets rid of a bunch of error-prone copy-paste and works in all cases.

#jira UE-198417
#rnx
#rb luke.thatcher

[CL 29014083 by mihnea balta in ue5-main branch]