Initial class lowering
This enables classes with no inheritance, simple constant member initializers, and blocks. The primary changes to the existing VClass system are the switch to VVar for mutable members, and the addition of a VProcedure for the blocks. The VVars should go away with codegen support for l-expressions, and the VProcedure should be extended to handle overridden and effectful initializers.

The next most immediately useful missing features are methods (deferred because they depend on `self`; this changelist just adds them to the class as top-level function constants) and effectful member initializers (including simpler ones like constructing new objects, which are in the BPVM; deferred to split out the archetype/subtype override logic from this changelist).

#rb yiliang.siew, saam.barati

[CL 29142435 by russell johnston in ue5-main branch]