IAS cache correctly handles failed data writes.
If a data write fails then we do not want to update our journal or add any items to maps, so instead we drop the "phrase" (a collection of cache items written together) and close it off. DiskJournal takes ownership of the phrase, but will simply discard it as

Note that this change includes test code. The edits to FDiskCache::ClosePhrase() are the actual fix.

#rb pj.kack,johan.berg
#rnx
#tests created a disk 64MiB disk and wrote cache data larger than it to force write errors. Uploaded a save data with a nearly full cache on platform S and browsed the locker until new code path was hit.

#ushell-cherrypick of 29175096 by Martin.Ridgers

[CL 29178588 by martin ridgers in ue5-main branch]