Add animation compression memory usage estimation
Compressing animations happens in the background when a cook is happening. The asset compilation manager has a maximum amount of memory that it is willing to give out and schedules work based on estimated memory usage. For animation compression the estimates can be wrong by an order of magnitude, which is too much to not introduce problems in the system.

The biggest factors are that memory usage can depend on the compression codec in use and on the number of codecs for an asset. The compression code tries all codecs in parallel. A particularly bad case is the RemoveLinearKeys codec, which has memory usage in the order of O(NumKeys * NumBones). This CL adds a new API for estimating the memory usage of running a compression codec and it adjusts the code that estimates the memory usage to actually account for running all codecs in parallel.

#rb nicholas.frechette

[CL 29268557 by sebastian schoner in ue5-main branch]