Reserve space for at least one entry at the end of the journal
In certain cases the remaining space for the journal might be only 16 bytes. In these cases we can only fit the phrase description and MaxEntries for the phrase willl be -1. This will cause an assert and over commit the journal. With this change we detect that case in the preceding phrase and add an identity entry which will be ignored when loading the journal.

#rb pj.kack, martin.ridgers

[CL 29271269 by johan berg in ue5-main branch]