Fixed TChunkedArray ending iteration prematurely
Prior to this change the begin/end() iterator objects would be compared by their the element pointer. The equivalent of "begin.elements + num != end.element". In cases were the number of elements was an exact multiple of a chunks count, "elements + num" would be an out-of-bounds address. It is never dereferenced so this is fine. However, depending on the allocator, one chunk could immediately follow another in memory such that "addr(B.elements + num) == addr(A.elements)". If A was the first chunk iterated, it would have the same element address as end() in B and iteration would finish prematurely.

#rb paul.chipchase
#rnx
#tests analysis of good/bad data from Horde, created a unit test to induce and prove scenario
#ushell-cherrypick of 29232307 by Martin.Ridgers

[CL 29283192 by martin ridgers in ue5-main branch]