Support writing partial buffers to the cache
Current caching mechanism only supports peeling off complete buffers from the memory cache to the cache file. This change adds support for writing partial buffers, with the remainder written in consecutive passes. Once the entire buffer is written to the file cache the journal entry will be added. This allows for greater utilization of the alloted write budget on certain platforms.

#rb martin.ridgers

[CL 29450620 by johan berg in ue5-main branch]