Skein / ConflictUE: Streamlined behavior when user clicks CHECK-IN CHANGES.
Previously, the SYNC was done prior to showing the SUBMIT FILES dialog. If another user submitted anything while the user was typing their changelist description, the CHECK IN would fail (for Skein) because the user is no longer checking in as latest. By moving the SYNC to after the changelist description is completed, this is avoided. In addition, this allows the user once again to click the CHECK IN CHANGES without syncing to head if he just wants to see what local changes he has that need checking in - a workflow which I imagine is not uncommon.

Additionally, the check-in should be aborted if the SYNC results in the user entering conflict resolution.

#rb manuel.lang, marco.anastasi
#rnx

[CL 29484477 by wouter burgers in ue5-main branch]