Fixed double free segfault.
When one IAS object is nearing completion we send out a request for the next one. If the current request has in fact finished, and requesting the next one fails, the current request will end up added to the "done list" twice. This results in said request getting freed twice when the done list is processed.

#rb pj
#rnx
#tests wrote tests to reliably reproduce use-after-free and implement and verify fix

#ushell-cherrypick of 29481582 by Martin.Ridgers

[CL 29486741 by martin ridgers in ue5-main branch]