RayTracing: Fix alpha channel in RT translucency and PrimaryRay debug view and add Holdout support
The alpha channel is now tracked like in other passes (it was not working consistently before). This applies both to RT translucency and PrimaryRay debug view.

A holdout flag is now propagated from the shader to the raygen shader so that holdouts can be tracked properly. This involved also flagging camera rays (as the holdout logic should only apply in these cases).

Also fix alpha channel of RT debug views for consistency

#jira UE-193122
#rb Sebastien.Hillaire

[CL 29551778 by chris kulla in ue5-main branch]