Removed check in ValidateLinkedPinTypes
Some users are hitting a check where no compiler relevant pin cannot be found during validation. It's unclear if this is an expected scenario. For now, we'll relax this to a soft-fail, and add logging to help track down the issue. If this turns out to be a valid scenario, we can remove the log in a future change.

#rb dan.oconnor

[CL 29602075 by dave jones2 in ue5-main branch]