Sequencer: fix more edge cases with camera cut track camera locking
When unlocking the camera cut track, we were only discarding the transient pre-animated state. This caused two problems:

1. This was mostly promoting saved state to "persistent" state, which means that we could still mess up the viewports when closing the Sequencer editor.
2. We have 2 things to do in pre-animated state: clear the viewport cinematic lock and optionally restore the viewport position. Since we always need to do the former, we need to always save pre-animated state, but be able to check for the latter while restoring.

So now we always save pre-animated state, check for the restoring of viewport position inside the restoring function, and completely discard all pre-animated state when unlocking cameras. This last bit required writing a new function on the pre-animated state linker extension because we need to discard all tracking and bookkeeping of contributors... (we don't want anything to happen once we scrub out of the camera cuts)

This change also fixes several methods of FPreAnimatedStateExtension not calling into the track instance and track instance input sources.

#jira UE-199794
#rb andrew.rodham

[CL 29620567 by ludovic chabant in ue5-main branch]