Sequencer- Interface for objects to receive events when they are bound/unbound by Sequencer.
Added a new interface that allows any UObject to implement an interface (either in C++ or BP) to receive events when Sequencer binds/unbounds that object. To get the events requires the sequence add a Binding Lifetime track to the binding. If the BindingLifetime track remains infinite in length, then the events will fire at the beginning and end of sequence playback respectively. If the binding lifetime track has a finite range, then the events will fire when entering/exiting the binding lifetime section.

[REVIEW] [at]ue-sequencer, [at]guillaume.guay
#jira UE-105285
#rb Max.Chen

[CL 29681701 by david bromberg in ue5-main branch]