SoftsEvolution -- rework PBDEvolution to be more group-oriented and setup to add force-based solver.
SoftsEvolution currently just has the same functionality of PBDEvolution, but it's structured in a way that will make it easier to add the force-based solver where we need to build and solve a linear system.

Main change is adding the concept of a ParticleRange which holds a pointer to the Particles SOA and an index and size to create an ArrayView. This works similar to the existing ParticleHandle, only it's for a contiguous range of particles.

The new Softs::FEvolution manages ranges, and the constraints have been updated to know how to act on ranges.

On the constraint side, the main change is the expectation that all internal indexing is working on local indices for the range rather than global indices. This makes things much simpler for data that only exists for a particular cloth object (range) like weight maps and triangle mesh indices.

Additionally, it is now explicitly enforced in Softs::FEvolution that each range corresponds with a single groupId. This simplifies collision code and some code related to managing transforms.

Finally, some more things were moved out of the Evolution and into constraint rules: per-substep initialization, external forces, and collision constraints. This removes a lot of the PropertyCollection management and weird group override settings in the cloth plugin. This also made it easier to refactor the initial calculation of P (euler stepping external forces and applying damping) to allow future additional ISPC or otherwise optimizations. Per-substep initialization was split into rules that are const on particles that can run in parallel and per-substep constraints that affect particles (e.g., long range constraint).

#rb kriss.gossart
#jira UE-200450

[CL 29751240 by alex mcadams in ue5-main branch]