Add support for weak maps to the Verse GC and implement a very basic weak map.
By weak maps, I mean: if a weak map is live, the map has a K->V entry, and K is live, then the map marks V, otherwise it doesn't.

Supporting this in a concurrent GC is hard because we can't "just" fixpoint the map in stop-the-world like everyone else does.

This change gets around that problem by enabling each cell to maintain a map->value mapping. This is done by hijacking a word in the libpas header for each page and using it to maintain a key->map->value map.

This also changes how native memory accounting works. The new way requires tracking live native bytes during marking. The benefit of doing it that way is that we can handle the native memory usage shrinking (for example, due to removal of an element from a TMap).

#rb jared.cotton

[CL 29754529 by phil pizlo in ue5-main branch]