Fixed a large number of objects being registered with TEDS after a script recompile.
Whenever scripts recompiles a large number of object come through TEDS' reinstancing manager. A previous change added code that reregisters actors, but this has the unexpected side effect of now also registering everything that it send over during the recompile. The fix is to only reregister objects that have a memento recorded for them. This implies that the object was previously seen in TEDS and that there is still an interest in it, meaning any other objects can be ignored.

#rb logan.buchy

[CL 29760024 by ronald koppers in ue5-main branch]