Fixed FNullWindow forcing windows to be zero-sized before it was properly initialized
SWindow may call AdjustCachedSize before calling Initialize during its initial creation. This would previously result in the window being forced to zero size and a subsequent call to Initialize would never reshape the Slate window.
This is now fixed by caching whether the window has been initialized or not. This is more in line with the other platform windows that generally do a noop if they have not created a window handle yet.

#jira UE-200363
#rb William.Belcher

[CL 29776117 by andrew rodham in ue5-main branch]