Automation: Source indexing: Fixed source indexing working incorrectly for branches using the 'import' functionality on P4.
On Windows the code assumed that all source files come from the game's branch, which may not be true when importing parts of other branches. Modified the code to query the locally synced files using 'p4 have' which returns a list of correct local - depot mappings. Minor but related fixes on other platforms.

#rb David.Harvey

[CL 29784146 by wojciech krywult in ue5-main branch]