Fixed HLSL->GLSL cross compile issues. Fixed missing FWaterMesSceneProxy::hbSupportsParallelGDME=false for GPU water quadtree path.
[Backout] - CL29753610
[FYI] tim.doerries
Original CL Desc
-----------------------------------------------------------------
[Backout] - CL29714981
[FYI] tim.doerries
Original CL Desc
-----------------------------------------------------------------
Implemented a path for generating and traversing the water quadtree on the GPU. Traversal creates indirect draw arguments which are used to draw the tessellated water mesh tiles. The quadtree is created by rasterizing water meshes into a supersampled and multisampled texture, downsampling that texture to quadtree LOD0 resolution and finally building a mip chain. Every frame when water needs to be rendered, all nodes (texels) in that quadtree texture are visited by a compute shader to determine which ones need to be drawn from a given view. The generated list of tiles is sorted into draw buckets (quad mesh tessellation and material) and finally a list of per-instance data is generated that is passed to the modified water vertex factory. The VF can now handle indirect draws with instance data read from a compacted buffer using another indirection to get to the correct offset.

The indirect draw pipeline also supports HZB based occlusion culling.

Reviewed in review 28664128

#rb Sebastien.Hillaire, jonathan.bard
#jira UE-141982
#rnx

[CL 29790252 by tim doerries in ue5-main branch]