Fix ClusterUnion child transforms for components added while asleep
We were relying on the cluster union to call MoveComponent on all components in the cluster, but this does not happen is the cluster is asleep or has not moved. So if a new componet was added (on the client), but the cluster does not wake or move (or OnRep_ChildToParent is called after it moved) the new component will have whatever transform it was spawned with, which is usually the world transfomr from the server. This meant that the visials and game thread collisions were not in the correct location which can caus eproblem for system that depend on it.

#rb cedric.caillaud

[CL 29835964 by chris caulfield in ue5-main branch]