Skip PAK file index encryption when -SkipEncryption is set
We have an option to skip encryption for PAK files but is it not consistently applied. This shows up for example when building the Content Worker: it sets the -SkipEncrpytion parameter in its packaging stage, yet the PAK tool is invoked with -encryptindex. This is relevant for modular builds, since those might not have encryption keys loaded at the time that the first PAK files are loaded.

[REVIEW] graeme.thornton
#rb Graeme.Thornton

[CL 29887972 by sebastian schoner in ue5-main branch]