Fix Cluster Union initial anchored state.
Problem fixed: The initial state of a cluster in non-anchored and simulating. On the server everything works fine if you set the anchored state after adding all components to the cluster and checking for support because it all happens on the same tick. But on the client we may receive the components and anchored state change over multiple frames. If anthing interacts with the client-side cluster after components are added but before the anchored flag replciates, it will move and later freeze in its new location when the anchored state is received.

#rb cedric.caillaud

[CL 29926620 by chris caulfield in ue5-main branch]