[Re-submit]  Material Translate:  Optimize out unused preshaders (meaning preshaders not referenced by any HLSL) in Translate_Legacy.  On a worst case shader, reduces preshaders by 4x, and parameter fetches and ops by 2.5x.  On a more typical shader, reduces preshaders and parameter fetches by 40%, and ops by 33%.
A scope class tracks when we are in a code path that's generating a uniform expression, and suppresses generation of preshaders for that case.  More detailed comments on FAddUniformExpressionScope class in the code.  Utility virtual function "GetChildren" added to FMaterialUniformExpression, which allows generalized traversal of a uniform expression tree, which I used for debug code (since removed), but others may find it useful.

Change was backed out -- resubmitting following CL 29887540, which appears to have fixed a bug that this CL triggered when checked in.

#rnx
#rb massimo.tristano

[CL 29926806 by jason hoerner in ue5-main branch]