Get rid of custom "Growth Factor" of AudioMixerSourceManager command buffer
UE's default growth factor is 1.35 for each re-allocation of the array: (so `Max_N = 1.35 * Max_N-1`), so _not_ doubling every time. 1.35 is a reasonable factor, and using a smaller value doesn't seem to gain that much benefit and adds code complexity.

The actual performance boost comes from "reserving" the array at creation. And preventing log spam avoids the frame hitches.

Also, avoid holding onto the RenderThreadIndex Lock for longer than we need to. Use a ThreadSafeCounter for "NumTimesOvergrown" as well so we avoid needing a lock for that as well.

#jira UE-200738
#rb aaron.mcleran
[FYI] jimmy.smith
#tests PIE

[CL 29997906 by jake burga in ue5-main branch]